pyproject.toml
[project]
name = "gym-delivery-bot"
version = "1.0.0"
description = "Telegram bot for gym food delivery"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }

authors = [
    { name = "Gym Delivery Team" }
]

dependencies = [
    "aiogram==3.10.0",

    "sqlalchemy==2.0.25",
    "asyncpg==0.29.0",
    "alembic==1.13.1",

    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",

    "structlog==24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",

    "black==24.1.1",
    "isort==5.13.2",
    "mypy==1.8.0",

    "pre-commit==3.6.0",
]

[project.scripts]
bot = "app.main:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true